Fix -c usage
