+++ ckan/cli/cli.py
@@ -3,14 +3,13 @@
 import logging
 from collections import defaultdict
 
-import ckan.plugins as p
 import click
 from ckan.cli import config_tool
 from ckan.cli import (
     jobs,
     datapusher,
     front_end_build,
-    click_config_option, db, load_config, search_index, server,
+    click_config_option, db, search_index, server,
     profile,
     asset,
     datastore,
@@ -27,7 +26,6 @@ from ckan.cli import (
     user
 )
 
-from ckan.config.middleware import make_app
 from ckan.cli import seed
 
 log = logging.getLogger(__name__)
@@ -37,7 +35,7 @@ class CustomGroup(click.Group):
     def get_command(self, ctx, name):
         cmd = super(CustomGroup, self).get_command(ctx, name)
         if not cmd:
-            ctx.invoke(self)
+            ctx.forward(self)
             cmd = super(CustomGroup, self).get_command(ctx, name)
         return cmd
 
@@ -60,23 +58,12 @@ class CustomGroup(click.Group):
                         formatter.write_dl(rows)
 
 
-class CkanCommand(object):
-
-    def __init__(self, conf=None):
-        self.config = load_config(conf)
-        self.app = make_app(self.config.global_conf, **self.config.local_conf)
-
-
 @click.group(cls=CustomGroup)
 @click.help_option(u'-h', u'--help')
 @click_config_option
-@click.pass_context
-def ckan(ctx, config, *args, **kwargs):
-    ctx.obj = CkanCommand(config)
-    for plugin in p.PluginImplementations(p.IClick):
-        for cmd in plugin.get_commands():
-            cmd._ckanext = plugin.name
-            ckan.add_command(cmd)
+# @click.pass_context
+def ckan(config, *args, **kwargs):
+    pass
 
 
 ckan.add_command(jobs.jobs)
